Refactor: Remove chalice_support and fastapi_support module and update imports to use chalice and fastapi directly
