+++ agave/chalice/lib/mongoengine/enum_field.py
