+++ agave/chalice/lib/mongoengine/model_helpers.py
