+++ agave/chalice/models/base.py
